% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fpop}
\alias{fpop}
\title{fpop with many cost functions}
\usage{
fpop(vectData, penalty, change = "mean")
}
\arguments{
\item{vectData}{vector of data to segment}

\item{penalty}{the penalty coefficient. A positive number}

\item{change}{A string defining the kind of cost function to use}
}
\value{
a list of two vectors and a value  = (changepoints, segment parameters, global cost).
'changepoints' is the vector of changepoints (we give the last element of each segment).
'parameters' is the vector of successive parameters of each segment
'globalCost' is a number equal to the global cost of the graph-constrained segmentation
}
\description{
A fast implementation of the fpop algorithm with many cost functions available
}
